FWA stream network

Get stream network features from gnis_name.

fwa_stream_gnis(gnis_name, columns = NULL, bbox = NULL, epsg = 4326)

Arguments

gnis_name
A character string of the gnis_name.

columns
A vector of strings of the column names to include. If NULL (default), all columns are retained.

bbox
A vector of numbers indicating bounding box to limit output features to, e.g. c(minLon, minLat, maxLon, maxLat).

epsg
A positive whole number of the epsg to transform features to.

Value

A sf object

Examples

fwa_stream_gnis('Sangan River')

#> Simple feature collection with 1 feature and 6 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -131.9972 ymin: 53.94994 xmax: -131.9169 ymax: 54.03317
#> geographic CRS: WGS 84
#>     id blue_line_key fwa_stream_networks_label_id    gnis_name stream_order
#> 1 9046     360879896                         9046 Sangan River            5
#>   watershed_group_code                       geometry
#> 1                 GRAI MULTILINESTRING ((-131.9185...